requirements.txt
pyspark==3.5.0
delta-spark==3.0.0
kafka-python==2.0.2
apache-airflow==2.8.0
apache-airflow-providers-databricks==6.0.0
apache-airflow-providers-snowflake==5.0.0
pytest==8.1.0
pandas==2.2.0
